refactor: relocate HttpProblemDetails to main me.desair.tus.server package
