+++ src/main/java/me/desair/tus/server/RequestHandler.java
@@ -2,7 +2,6 @@ package me.desair.tus.server;
 
 import java.io.IOException;
 import me.desair.tus.server.exception.TusException;
-import me.desair.tus.server.rufh.HttpProblemDetails;
 import me.desair.tus.server.upload.UploadLockingService;
 import me.desair.tus.server.upload.UploadStorageService;
 import me.desair.tus.server.util.TusServletRequest;
