+++ src/main/java/me/desair/tus/server/rufh/handler/RufhCreationPostRequestHandler.java
@@ -6,8 +6,8 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import me.desair.tus.server.HttpHeader;
 import me.desair.tus.server.HttpMethod;
+import me.desair.tus.server.HttpProblemDetails;
 import me.desair.tus.server.exception.TusException;
-import me.desair.tus.server.rufh.HttpProblemDetails;
 import me.desair.tus.server.rufh.InterimResponseStrategy;
 import me.desair.tus.server.upload.UploadInfo;
 import me.desair.tus.server.upload.UploadLockingService;
