+++ src/main/java/me/desair/tus/server/rufh/handler/RufhDeleteRequestHandler.java
@@ -2,8 +2,8 @@ package me.desair.tus.server.rufh.handler;
 
 import java.io.IOException;
 import me.desair.tus.server.HttpMethod;
+import me.desair.tus.server.HttpProblemDetails;
 import me.desair.tus.server.exception.TusException;
-import me.desair.tus.server.rufh.HttpProblemDetails;
 import me.desair.tus.server.upload.UploadInfo;
 import me.desair.tus.server.upload.UploadLockingService;
 import me.desair.tus.server.upload.UploadStorageService;
