+++ src/main/java/me/desair/tus/server/util/AbstractTusExtension.java
@@ -5,12 +5,12 @@ import java.io.IOException;
 import java.util.LinkedList;
 import java.util.List;
 import me.desair.tus.server.HttpMethod;
+import me.desair.tus.server.HttpProblemDetails;
 import me.desair.tus.server.ProtocolVersion;
 import me.desair.tus.server.RequestHandler;
 import me.desair.tus.server.RequestValidator;
 import me.desair.tus.server.TusExtension;
 import me.desair.tus.server.exception.TusException;
-import me.desair.tus.server.rufh.HttpProblemDetails;
 import me.desair.tus.server.upload.UploadLockingService;
 import me.desair.tus.server.upload.UploadStorageService;
